src/evaluation/pyproject.toml
[project]
name = "evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "accelerate>=1.10.1",
    "beautifulsoup4>=4.14.2",
    "bs4>=0.0.2",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "langchain>=1.0.1",
    "langchain-community>=0.4",
    "langchain-ollama>=0.2.0",
    "langchain-text-splitters>=1.0.0",
    "line-bot-sdk>=3.21.0",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.10",
    "pypdf>=6.1.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
cpu = [
    "torch>=2.8.0",
    "torchvision>=0.20.0",
]

rocm = [
    "torch>=2.8.0",
    "torchvision>=0.20.0",
    "pytorch-triton-rocm>=2.1.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" }, # Only install with --extra cpu
  { index = "pytorch-rocm", extra = "rocm" }, # Only install with --extra rocm
]

torchvision = [
  { index = "pytorch-cpu", extra = "cpu" }, # Only install with --extra cpu
  { index = "pytorch-rocm", extra = "rocm" }, # Only install with --extra rocm
]

pytorch-triton-rocm = [
  { index = "pytorch-rocm", extra = "rocm" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm6.4"
explicit = true

[tool.uv]
conflicts = [
    [
      { extra = "cpu" },
      { extra = "rocm" },
    ],
]
